Added Feats, begin work on Combat tab
Added support for Feats, including using the new ItemListDisplay to display the feats! I began work on the Combat tab, but there's still a decent way to go. I created a DefenseEditor control, but it's highly likely I'm going to just get rid of it or rename it to ArmorEditor.