Add Combat tab
Set up the Combat tab's UI. The backend hookups/mapping and the JSON file mapping is not done yet, that'll be done... shortly? Lol. Removed the DefenseEditor as I ended up not using that. Also added a DetailsButton control that'll be useful to just quickly throw in where I need it.

I'll admit I'm a bit concerned that this main window will end up with just as many elements as the PathfinderJson's main window, but I guess we'll find out. Big difference is there won't be quite as many various properties/variables in Sentinels as there are in Pathfinder (as Sentinels is a much more modular system, with characters being customized/augmented via attributes rather than by changing existing mechanics in the system like Pathfinder does), so hopefully this main window won't end up as bad.