Add Combat values/calculations, Power stat handling
Add calculations for the Combat tab. Add values in backend for the SentinelsSheet to store needed data. By default, calculations will be automatically updated whenever the user's level, ability score, or combat backend values change. Set up the Update Calculations method and the Calculate menu. Also, modifiers are now stored in a dictionary in the backend for easy access. Also added handling and a stored string for changing/accessing the power stat.