Add NewSheet dialog, updates to JSON formatting
Added in a NewSheet dialog, based upon the design present in PathfinderJSON. Also made updates to JSON formatting, defining the names certain values have. SentinelsJson saves its levels as a number (unlike PathfinderJson, where the "level" is a string with the character's class and level), so to keep them from conflicting, levels in SentinelsJSON are referred to as "charLevel" in JSON. In this way, SentinelsJSON can continue to open PathfinderJSON files (for the sake of copying over UserData as needed). I may change this in the future to create a generic "UserDataSheet" class so I don't need to worry about conflicting names/types between systems.