Finish CreateSentinelsSheet function, add Saves
Finish the CreateSentinelsSheet function, which the program uses to generate the sheet based upon the contents of the editors. This is used to save files, as well as swap between the editor views and JSON view. With this, syncing the views was enabled. Added editor for saves (i.e. Reflex save, Will save) on the General tab.