Cleaned up FeatEditor, SelectableListItem
With these changes made, it'll be easier for me to begin making more editors (and replace the editors in PathfinderJson). I've tried to combine as much functionality as I can into existing controls or into the base SelectableListItem control, so that I don't have to worry about copying it over into each new editor. Surprisingly, I'm able to reference event handlers from the base class in XAML, although the designer doesn't seem to like to support that.