Add Skills tab, SkillEditor class
Skills tab has been added to the MainWindow, but is currently just in testing right now. Later the SkillList class will be used to dynamically make the list of all the skills. Also added in the SkillEditor, which has a different design than was used in PathfinderJSON. This new design will be ported over to PathfinderJSON eventually as well.