Add Skills tab, skill handling
Added in the Skills tab. Skills are loaded in based upon the skill list that that particular sheet uses. In the future, I'll add in support to load in custom skill lists from a file (and then, afterwards, a skill list editor). Skills are loaded into a list of SkillEditors on the Skills tab, each one controlled by dependency properties and binding. Includes seamless support for changing the relevant ability for a given skill; this change is saved in the sheet file. To avoid conflict and potential issues with PathfinderJSON files, skills in these JSON files are called "charSkills".