Add Sheet Settings dialog, fix Pathfinder skill list
This dialog will allow users to change some (otherwise hidden) values about their sheets, such as character points per level and the skill list to use. Enhanced support for loading skill lists from files. The sheet settings dialog will also display the SheetSettings dictionary, but with the warning that these are advanced values that you should be careful with. Fixed the pathfinder skill list to have the appropriate modifiers. Updated SSUI to include a SelectionChanged event for FileSelect and to change the constructor for SelectableItem. Added a "None" skill list option to the New Sheet dialog.