Improvements to ILD
Made readability improvements and expansion to ILD, such as renaming the function GenerateMappedProperties to GetAllPropertyValues. Added more functionality for the filtering code.

From this point on, I'll be moving the work on ILD to PathfinderJson, as I have a more complete project there which I can finish building ILD with.